api/requirements.txt
aiohttp==3.8.3
aioresponses==0.7.4
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
argcomplete==3.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.6.2
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
cryptography==40.0.2
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
Flask==2.3.2
fqdn==1.5.1
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
holehe==1.61
httpcore==0.16.3
httpretty==1.1.4
httpx==0.23.3
idna==3.4
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyter_server==2.0.6
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
leakcheck==1.0.0
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.4
mysql-connector-python==8.0.33
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
notebook_shim==0.2.2
numpy==1.24.2
openpyxl==3.0.10
outcome==1.2.0
packaging==22.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pipx==1.2.0
platformdirs==2.6.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.15.0
Pygments==2.14.0
PyJWT==2.7.0
pyOpenSSL==23.1.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.4
python-Wappalyzer==0.3.1
pytz==2022.7.1
PyYAML==6.0
pyzmq==24.0.1
requests==2.31.0
requests-futures==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
scrape-search-engine==0.2.2
selenium==4.9.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spark-parser==1.8.9
stack-data==0.6.2
stem==1.8.1
termcolor==2.2.0
terminado==0.17.1
terminaltables==3.1.10
tinycss2==1.2.1
tornado==6.2
torrequest==0.1.0
tqdm==4.64.1
traitlets==5.8.0
trio==0.22.0
trio-websocket==0.10.2
uncompyle6==3.9.0
uri-template==1.2.0
urllib3==2.0.2
userpath==1.8.0
wcwidth==0.2.5
webcolors==1.12
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.3.4
wsproto==1.2.0
xdis==6.0.5
yarl==1.8.2
